as-added repo and entity file
