+++ src/main/java/edu/ucsb/cs156/example/repositories/UCSBDiningCommonsMenuItemRepository.java
@@ -1,29 +1,12 @@
+package edu.ucsb.cs156.example.repositories;
 
-package edu.ucsb.cs156.example.entities;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+import edu.ucsb.cs156.example.entities.UCSBDiningCommonsMenuItem;
 
-import java.time.LocalDateTime;
+@Repository
+public interface UCSBDiningCommonsMenuItemRepository extends CrudRepository<UCSBDiningCommonsMenuItem, Long> {
+           
 
-import javax.persistence.Entity;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.GeneratedValue;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Entity(name = "UCSBmenuitems")
-public class UCSBDiningCommonsMenuItem {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private long id;
-
-  private String diningCommonsCode;
-  private String name;  
-  private LocalDateTime station;
+    
 }
